Use a MS attribute to tag methods.
LoaderOptimizationAttribute may be set to 'none' and beyond that,
it is ignored on all methods that aren't the main method of an
executable file. Beyond that, it shares a weak syntactical bond
to what I'm using it for, that is to optimize my workflow for the
Unity Mod Manager Loader.

(I know, it's a stretch. But it works, and doesn't need PBUtils
references on everything!)